// src/types.ts
export interface CustomTheme {
  id: string;
  name: string;
  primaryColor: string; // Nền chính
  secondaryColor: string; // Chữ chính
  accentColor: string; // Màu nhấn
  successColor: string; // Thành công
  warningColor: string; // Cảnh báo
  errorColor: string; // Lỗi
  mutedColor: string; // Mờ
  isDark: boolean;      // Nền tối hay sáng
}